Billmate Plugin Version: 1.29 fix: double invoice generation in cardpay and bank fix: wrong typecasting of secret key, make incorrect key due to casting

+++ system/library/billmate/OpenCartBillmate.php
@@ -23,7 +23,7 @@
 		$ssl = true;
 		$debug = false;
 		$eid = (int)$eid;
-		$key = (float)$key;
+		$key = (int)$key;
 
         $k = new BillMate( $eid ,$key,$ssl,$debug);
